Pullup ticket #2706 - requested by adrianp pdksh: bug fix
Revisions pulled up:
- shells/pdksh/Makefile				1.18
- shells/pdksh/files/c_ulimit.c			1.4
- shells/pdksh/files/lex.c			1.4
---
Module Name:	pkgsrc
Committed By:	tnn
Date:		Sat Feb 21 20:06:30 UTC 2009

Modified Files:
	pkgsrc/shells/pdksh: Makefile
	pkgsrc/shells/pdksh/files: c_ulimit.c lex.c

Log Message:
Merge the following revisions from NetBSD src:
c_ulimit.c 1.9: avoid sign extension problem
lex.c 1.13: bugfix related to nested quotes
Bump PKGREVISION.

+++ shells/pdksh/files/c_ulimit.c
@@ -1,4 +1,4 @@
-/*	$NetBSD: c_ulimit.c,v 1.3 2008/06/15 14:20:08 tnn Exp $	*/
+/*	$NetBSD: c_ulimit.c,v 1.3.10.1 2009/02/23 13:18:56 tron Exp $	*/
 
 /*
 	ulimit -- handle "ulimit" builtin
@@ -187,7 +187,7 @@ c_ulimit(wp)
 			    bi_errorf("invalid limit: %s", wp[0]);
 			    return 1;
 			}
-			val = rval * l->factor;
+			val = (unsigned long)rval * l->factor;
 		}
 	}
 	if (all) {
